requirements.txt
attrs==23.2.0
colorzero==2.0
contourpy==1.2.1
cycler==0.12.1
dnspython==2.7.0
fonttools==4.53.1
gpiozero==2.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.0
packaging==24.1
pigpio==1.78
pillow==10.3.0
pycryptodome==3.20.0
pymongo==4.12.1
pyparsing==3.1.2
pyproject-toml==0.0.10
python-dateutil==2.9.0.post0
referencing==0.35.1
rpds-py==0.19.0
scipy==1.14.0
setuptools==70.2.0
six==1.16.0
toml==0.10.2
wheel==0.43.0
